# Use make size hint when copying channels in Nick()