@fluffle 2023-11-17
How many channels is your bot in? :-D